feat: move CA to hero section below browse icons button
